requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
DateTime==5.2
defusedxml==0.7.1
dj-database-url==2.1.0
Django==3.2.22
django-address==0.2.8
django-allauth==0.57.0
django-decouple==2.1
django-environ==0.11.2
# django-heroku==0.3.1
geographiclib==2.0
geopy==2.4.0
gunicorn==21.2.0
idna==3.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
psycopg2-binary
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.6
whitenoise==6.5.0
zope.interface==6.1
